requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.15
MarkupSafe==3.0.3
python-engineio==4.12.3
python-socketio==5.14.2
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
